requirements.txt
# Core LangChain and Ollama
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.0.1
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0

# Optional: OpenAI (not required for local-only)
# openai>=1.12.0

# Vector Database
faiss-cpu>=1.7.4

# Document Processing
python-dotenv>=1.0.0
tiktoken>=0.7.0
pypdf>=4.2.0

# Scientific Computing
numpy>=1.26.0
matplotlib>=3.8.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
datasets>=2.16.1

# Production API (optional - for production deployment)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
